+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-
+* Factorize and clean up code from: `FlaneurImageDescription`
+    * Using enum to associate `sourceType` and its associated value (ie URL, asset or image)
+    * No caching done by `FlaneurImageDescription` for `PHAsset` as it's not its role and was not reliable (no track of the fetch options like quality, size, etc.)
+    * Stop conforming to `NSObject`
+* Added `FlaneurImageView` as a view designed to be reusable (ie a subview of `UITableViewCell` or `UICollectionReusableView`)
 
 ## [0.1.0] - 2017-10-24